AlphaBeta Pruning works? Maybe.
AlphaBeta pruning has been written. The heuristic is absolute trash at this moment. Also there's a todo in AI.pde on line 105 and 135. I'm confused whether I should return the value, or the entire ABObj. To be determined. I think I'm giving up on online. Maybe I'll implement it in the future, but probably not for the IA submission. I'm just too busy.