Success! ABP works like it should.
It works! I can even increase the depth. The only problem becomes computing resources. If I increase depth to 3, I get an error that says I'm out of memory. I tried to initiate garbage collection, but I'm not sure if there is more that I can do. Also the CPU usage shoots up to 97% on an overclocked i5 4690k. Crazy! The heuristic is still garbage, but it shouldn't be hard to improve it.