My thoughts were wrong.
I deleted the old heuristic comment.

But, in addition, the whole idea of a counter counter (tallied up streaks) failed. It looks like the heuristic I wanted to implement is dependent on finding win states, and needs at least a depth of seven (looks at 7 moves). This is currently infeasable because of memory expense. I should look into reducing memory down in the ABTree's ABNodes. I should find a way to measure the memory expense to do this. Things like ints can switched to bytes. Also during all of this, I need to figure out the garbage collection strategy.

After memory expense is taken care of, I need to build a win detector within the AI that marks a node as terminal.