Added child deletion to free memory.
Children are deleted when they have been heuristically evaluated, as long as the moves within them are not needed.

It's strange. I thought I cracked the memory problem, but I increased the depth to 3, and when I was one move away from a capture move, it ran out of memory. I was able to do a depth of 4 without a scenario such as that going on (but cpu usage was suprisingly low during that). I should look into memory usage mid-search.

The other thing I'm considering is that I'm using an ArrayList. Is that really better than an array for memory consumption? I don't know.

The GameAI constructor was also changed to add depth. This helps out with the child deletion.