This is about as good as ABP is gonna get.
After a lengthy optimization process, I've determined that this is about the best it will ever be. I should try to implement a killer heuristic, but aside from that, there's really not much more optimization I can do.

Turns out that ABP really was not the best search tree algorithm for the task. I should have done a monte-carlo tree search instead. But, for now I'm gonna leave it as ABP. I'll implement the good heuristic, killer heuristics, and maybe--just maybe--multithreading. I'm not gonna get above a depth of three. That's for sure.

Down the line I'm interested in pursuing the MCTS method, but I should just focus on drafing the write up after I do the stuff above.