I finally figured out why it was dumb.
Well I guess it wasn't dumb. I was (because I made it). When I added the move randomization it interfered with the way that ABP Search prunes branches. It is specifically designed to look at the front if there are equal moves for the root's children. I also had to add an else statement to deal with the default array value being zero.

Things to do:
1. Second stage of the recursive tree-less setup
2. implementation of good heuristic
3. pre-sorting moves to both add randomness and speed up the computational speed of the move generation.