Got rid of memory leak
In the whole tessCheck() thing it was creating new children, clogging up the memory with stuff that is just needed for a quick analysis. I switched over to a temp variable, and it seems to be working (java garbage collection takes care of it).

Also got rid of some legacy stuff and fixed mislabeled methods.