I've done a lot, and it now plays well.
I just don't think 19x19 is feasible on a consumer computer. It would need a powerful server or something to run. Maybe a GPU even.

Even if it is possible (which I bet it is) it would require way more than I'm currently planning on doing. Something like a neural network could probably do it. But, for now, I'm gonna limit it to 13x13. It would be interesting to extend in the future.